Version 2/risk-agent/requirements.txt
langchain==0.1.20
langchain-openai==0.1.8
openai==1.35.0
fastapi==0.111.0
uvicorn==0.30.0
redis==5.0.6
supabase==2.5.0
pydantic==2.7.4
python-dotenv==1.0.1
pytest==8.2.2
pytest-asyncio==0.23.7
